Check if the directory for the exporting is specified, otherwise no export actions will be taken.

+++ TensorFlowTraining/export_environment.py
@@ -21,17 +21,20 @@ def init_export_environment(prm: ExportParameters):
     Keyword arguments:
     prm     -- the export parameters
     """
+    # Check if the export directory is specified
+    if (not prm.output_directory):
+        print('Warning: export output directory is not specified so nothing will be exported')
+        return;
     # Set the configuration for Google Colab
     if ('google.colab' in sys.modules and cfg_data_on_drive):
         if (not os.path.exists('/mnt/MyDrive')):
             print('Mounting the GDrive')
             from google.colab import drive
             drive.mount('/mnt')
-
         # Check the existence of the checkpoints directory
         gdrive_dir = os.path.join('/mnt', 'MyDrive', prm.trained_checkpoint_dir)
         if (not os.path.isdir(gdrive_dir)):
-            raise Exception('Error!!! The trained checkpoint dir doesn`t exist')
+            raise Exception('Error!!! The trained checkpoint directory doesn`t exist')
         if (os.path.exists('/content/trained-model')):
             os.unlink('/content/trained-model')
         os.symlink(gdrive_dir, '/content/trained-model', True)
@@ -47,8 +50,7 @@ def init_export_environment(prm: ExportParameters):
         if (os.path.exists('/content/exported-model')):
             os.unlink('/content/exported-model')
         os.symlink(gdrive_dir, '/content/exported-model', True)
-        gdrive_dir = os.path.join(prm.output_directory, 'exported-model')
-        print(f"Google drive's {gdrive_dir} is linked to /content/exported-model")
+        print(f"Google drive's {prm.output_directory} is linked to /content/exported-model")
         prm.output_directory = '/content/exported-model'
     else:
         if (not os.path.isdir(prm.trained_checkpoint_dir)):
@@ -61,7 +63,10 @@ def init_export_environment(prm: ExportParameters):
             raise Exception("Error: export directory cannot be the train directory")
         print(f'The trained model will be in {str(Path(prm.model_dir).resolve())}')
     # Copy the label file in the export directory
-    shutil.copy2(os.path.join(prm.trained_checkpoint_dir, 'label_map.pbtxt'), prm.output_directory)
+    try:
+        shutil.copy2(os.path.join(prm.trained_checkpoint_dir, 'label_map.pbtxt'), prm.output_directory)
+    except:
+        print(f"Warning: the file {os.path.join(prm.trained_checkpoint_dir, 'label_map.pbtxt')} doesn't exist and will not be copied to the output")
 
 if __name__ == '__main__':
     prm = ('prm' in locals() and isinstance(prm, ExportParameters) and prm) or ExportParameters.default
